requirements.txt
ago==0.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
astor==0.8.1
attrs==25.3.0
Automat==25.4.16
banks==2.2.0
beautifulsoup4==4.14.2
blake3==1.0.8
boto3==1.40.42
botocore==1.40.42
bs4==0.0.2
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
compressed-tensors==0.11.0
constantly==23.10.4
cryptography==46.0.2
cssselect==1.3.0
cupy-cuda12x==13.6.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.3.1
depyf==0.19.0
dill==0.4.0
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
dotmap==1.3.30
einops==0.8.1
elastic-transport==9.1.0
elasticsearch==9.1.1
email-validator==2.3.0
fastapi==0.118.2
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fastrlock==0.8.3
faust-cchardet==2.1.19
feedparser==6.0.12
filelock==3.20.0
filetype==1.2.0
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2024.6.1
gguf==0.17.1
greenlet==3.2.4
griffe==1.15.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hjson==3.1.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.1
hurry.filesize==0.9
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
interegular==0.3.3
itemadapter==0.12.2
itemloaders==1.3.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langdetect==1.0.9
lark==1.2.2
llama-index-core==0.14.8
llama-index-instrumentation==0.4.2
llama-index-workflows==2.11.1
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
lxml==6.0.2
lxml_html_clean==0.4.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
mistral_common==1.8.5
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.3
news-please==1.6.16
newspaper4k==0.9.3.1
ninja==1.13.0
nltk==3.9.2
numba==0.61.2
numpy==2.1.2
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.9.79
ollama==0.6.1
openai==2.8.0
openai-harmony==0.0.8
opencv-python-headless==4.12.0.88
outlines_core==0.2.11
packaging==25.0
pandas==2.3.3
parsel==1.10.0
partial-json-parser==0.2.1.1.post6
pillow==11.0.0
plac==1.4.5
platformdirs==4.5.0
portalocker==3.2.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.4.1
Protego==0.5.0
protobuf==6.33.1
psutil==7.1.3
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.0
pydantic-extra-types==2.10.6
pydantic_core==2.41.1
PyDispatcher==2.0.7
Pygments==2.19.2
pymongo==4.15.3
PyMySQL==1.1.2
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
qdrant-client==1.15.1
queuelib==1.8.0
ray==2.51.1
readability-lxml==0.8.4.1
redis==6.4.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-file==2.1.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.0
rpds-py==0.28.0
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
Scrapy==2.13.3
sentence-transformers==5.1.2
sentencepiece==0.2.1
sentry-sdk==2.41.0
service-identity==24.2.0
setproctitle==1.3.7
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tldextract==5.3.0
tokenizers==0.22.1
torch==2.8.0+cu129
torchaudio==2.8.0
torchvision==0.23.0+cu129
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
Twisted==25.5.0
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
w3lib==2.3.1
warcio==1.7.5
watchfiles==1.1.0
websockets==15.0.1
wrapt==2.0.1
xformers==0.0.32.post1
xgrammar==0.1.25
yarl==1.22.0
zope.interface==8.0.1
